requirements.txt
fastapi>=0.100.1
uvicorn[standard]>=0.23.2
httpx
asyncio
authlib
python-dotenv
pydantic
langchain